move get_ivar
